queueのテストと、Transactionのテストを分けた GitHub Action上だとTimeout時間の5sでテストが完了しないので、
``jest.setTimeout(10000)`` 10秒

を追加
